requirements.txt
python-wordpress-xmlrpc
toml
requests